requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
colorama==0.4.6
Django==4.1.0
django-autoslug==1.9.9
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.6
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
korrektor-py==1.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.9
pytest==7.3.2
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
